Rename the submodule
