+++ src/covvfit/_quasimultinomial.py
