+++ tests/test_quasimultinomial.py
@@ -1,4 +1,4 @@
-import covvfit._frequentist_jax as fj
+import covvfit._quasimultinomial as qm
 import jax
 import numpy.testing as npt
 import pytest
@@ -13,17 +13,17 @@ def test_parameter_conversions_1(seed: int, n_cities: int, n_variants: int) -> N
     growth_rel = jax.random.uniform(key1, shape=(n_variants - 1,))
     midpoint_rel = jax.random.uniform(key2, shape=(n_cities, n_variants - 1))
 
-    theta = fj.construct_theta(
+    theta = qm.construct_theta(
         relative_growths=growth_rel,
         relative_midpoints=midpoint_rel,
     )
 
     npt.assert_allclose(
-        fj.get_relative_growths(theta, n_variants=n_variants),
+        qm.get_relative_growths(theta, n_variants=n_variants),
         growth_rel,
     )
     npt.assert_allclose(
-        fj.get_relative_midpoints(theta, n_variants=n_variants),
+        qm.get_relative_midpoints(theta, n_variants=n_variants),
         midpoint_rel,
     )
 
@@ -36,11 +36,11 @@ def test_parameter_conversions_2(seed: int, n_cities: int, n_variants: int) -> N
         jax.random.PRNGKey(seed), shape=(n_cities * (n_variants - 1) + n_variants - 1,)
     )
 
-    growth_rel = fj.get_relative_growths(theta, n_variants=n_variants)
-    midpoint_rel = fj.get_relative_midpoints(theta, n_variants=n_variants)
+    growth_rel = qm.get_relative_growths(theta, n_variants=n_variants)
+    midpoint_rel = qm.get_relative_midpoints(theta, n_variants=n_variants)
 
     npt.assert_allclose(
-        fj.construct_theta(
+        qm.construct_theta(
             relative_growths=growth_rel,
             relative_midpoints=midpoint_rel,
         ),
